requirements.txt
pymongo==4.7.1 
python-dotenv==1.0.1 
rdflib==7.0.0